% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-dist.R
\name{bbox_dist}
\alias{bbox_dist}
\alias{st_xdist}
\alias{st_ydist}
\alias{st_xdist.sf}
\alias{st_xdist.sfc}
\alias{st_xdist.bbox}
\alias{st_ydist.sf}
\alias{st_ydist.sfc}
\alias{st_ydist.bbox}
\title{Get bounding box dimensions}
\usage{
st_xdist(x)

st_ydist(x)

\method{st_xdist}{sf}(x)

\method{st_xdist}{sfc}(x)

\method{st_xdist}{bbox}(x)

\method{st_ydist}{sf}(x)

\method{st_ydist}{sfc}(x)

\method{st_ydist}{bbox}(x)
}
\arguments{
\item{x}{(bbox/sfc/sf) Either a bounding box or a geometry.}
}
\description{
Get bounding box dimensions
}
