% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-lims.R
\name{bbox_lims}
\alias{bbox_lims}
\alias{st_xlim}
\alias{st_xlim.sf}
\alias{st_xlim.sfc}
\alias{st_xlim.bbox}
\alias{st_ylim}
\alias{st_ylim.sf}
\alias{st_ylim.sfc}
\alias{st_ylim.bbox}
\title{Get one-set of the bounding box}
\usage{
st_xlim(x)

\method{st_xlim}{sf}(x)

\method{st_xlim}{sfc}(x)

\method{st_xlim}{bbox}(x)

st_ylim(x)

\method{st_ylim}{sf}(x)

\method{st_ylim}{sfc}(x)

\method{st_ylim}{bbox}(x)
}
\arguments{
\item{x}{(bbox/sfc/sf) Either a bounding box or a geometry.}
}
\description{
Return either the x-axis limits
or y-axis limits.
}
\examples{
library(sf)
data(states_map)

x <- st_xlim(states_map)
x

# The crs is preserved
st_crs(x)
}
