% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sf}
\alias{read_sf}
\title{read simple features from file}
\usage{
read_sf(f)
}
\description{
read simple features from file, using GDAL and rgdal2
}
\examples{
f = system.file("example-data/continents", package = "rgdal2")
x = read_sf(f)
x
}

