% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_df}
\alias{sf_df}
\title{create sl_df object}
\usage{
sf_df(df)
}
\description{
create sl_df, which extends data.frame-like objects with a simple feature list column
}
\examples{
pt1 = c(0,1)
class(pt1) = "Point"
pt2 = c(1,1)
class(pt2) = "Point"
sf(list(pt1, pt2))
d = data.frame(a = 1:2)
d$geom = sfc(list(pt1, pt2))
df = sf_df(d)
d$geom2 = sfc(list(pt1, pt2))
sf_df(df) # warns
}

