% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-density.R
\name{st_density}
\alias{st_density}
\alias{st_density.sfc}
\alias{st_density.sf}
\title{Compute 2D density}
\usage{
st_density(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = 200,
  bins = NULL,
  truncate = TRUE,
  x_expansion = NULL,
  y_expansion = NULL,
  levels_low = NULL,
  levels_high = NULL
)

\method{st_density}{sfc}(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = NULL,
  bins = NULL,
  truncate = TRUE,
  x_expansion = NULL,
  y_expansion = NULL,
  levels_low = NULL,
  levels_high = NULL
)

\method{st_density}{sf}(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = NULL,
  bins = NULL,
  truncate = TRUE,
  x_expansion = NULL,
  y_expansion = NULL,
  levels_low = NULL,
  levels_high = NULL
)
}
\arguments{
\item{x}{(sf/sfc) [missing] Spatial data}

\item{return_geometry}{[\code{"point"}] (character) The geometry type to return.
Options: \code{"point"}, \code{"raster"}, \code{"contour"}/\code{"isoline"},
\code{"polygon"}/\code{"isoband"}.}

\item{method}{(character) [\code{"kde2d"}] How should density be computed?}

\item{bw}{(numeric) [\code{NULL}] Binwidth}

\item{n}{(numeric) [\code{NULL}] Grid size}

\item{bins}{(numeric) [\code{NULL}] Number of contour bins}

\item{truncate}{(logical) [\code{TRUE}] See \code{KernSmooth::bkde2D}}

\item{x_expansion, y_expansion}{(numeric) [\code{NULL}] Expansion multiple
applied to x-/y-range and used in \code{lims} and \code{range.x}
arguments for methods \code{"kde2d"} and \code{"bkde2D"}, respectively.}

\item{levels_high, levels_low}{(numeric) Low/high breakpoints
for isolines and isobands.}
}
\description{
Choose from methods \code{"kde2d"} and \code{"bkde2D"},
and various return geometries
}
