% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-extent.R
\name{st_extent}
\alias{st_extent}
\title{Return the bounding box as 'sf' or 'sfc'}
\usage{
st_extent(x, type = "sf")
}
\arguments{
\item{x}{(sf/sfc) Spatial data}

\item{type}{(character) Return type; either \code{"sf"}
(Default) or \code{"sfc"}}
}
\description{
Return the bounding box as 'sf' or 'sfc'
}
