% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{example-data}
\alias{example-data}
\alias{ngp}
\alias{states_map}
\title{Example data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33642 rows and 6 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 63 rows and 3 columns.
}
\usage{
ngp

states_map
}
\description{
\code{ngp} contains data for U.S. Natural
Gas Pipelines minus Alaska (2018-04-01) courtesy of
Homeland Infrastructure Foundation-Level
Data (HIFLD).

\code{states_map} is a map of the lower-48
United States of America.
}
\seealso{
For more about the natural gas pipeline
data, please see:
https://www.arcgis.com/sharing/rest/content/items/f44e00fce8b943f69a40a2324cf49dfd/info/metadata/metadata.xml?format=default&output=html
}
\keyword{datasets}
