% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-any.R
\name{st_any}
\alias{st_any}
\title{Check for minimum and maximum occurences of TRUE}
\usage{
st_any(x, at_least = 1L, at_most = Inf)
}
\arguments{
\item{x}{(sgbp/matrix) Return object of a simple feature
binary logical function.}

\item{at_least}{(integer) Minimal occurences of \code{TRUE}.
Default is \code{1L}.}

\item{at_most}{(integer) Maximum occurences of \code{TRUE}.
Default is \code{Inf}.}
}
\description{
Check for minimum and maximum occurences of TRUE
}
