% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-density.R
\name{st_density}
\alias{st_density}
\alias{st_density.sfc}
\alias{st_density.sf}
\title{Compute 2D density}
\usage{
st_density(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = 200,
  bins = NULL,
  range.x = NULL,
  truncate = TRUE
)

\method{st_density}{sfc}(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = NULL,
  bins = NULL,
  range.x = NULL,
  truncate = TRUE,
  levels_low = NULL,
  levels_high = NULL
)

\method{st_density}{sf}(
  x,
  return_geometry = "point",
  method = "kde2d",
  bw = NULL,
  n = NULL,
  bins = NULL,
  range.x = NULL,
  truncate = TRUE,
  levels_low = NULL,
  levels_high = NULL
)
}
\arguments{
\item{x}{(sf/sfc) Spatial data}

\item{return_geometry}{(character) What gets returned?}

\item{method}{(character) How should density be computed?}

\item{bw}{(numeric) Binwidth}

\item{n}{(numeric) Grid size}

\item{bins}{(numeric) Number of contour bins}

\item{range.x}{(numeric-list) See \code{KernSmooth::bkde2D}}

\item{truncate}{(logical) See \code{KernSmooth::bkde2D}}
}
\description{
Choose from methods \code{"kde2d"} and \code{"bkde2D"},
and various return geometries
}
