% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-extent.R
\name{st_extent}
\alias{st_extent}
\title{Return the bounding box as 'sf' or 'sfc'}
\usage{
st_extent(x, type = "sf")
}
\arguments{
\item{x}{(sf/sfc) Spatial data}

\item{type}{(character) Return type; either \code{"sf"}
(Default) or \code{"sfc"}}
}
\description{
Return the bounding box as 'sf' or 'sfc'
}
\examples{

suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(ggplot2))
data(states_map)

tx_extent <- states_map \%>\%
  filter(region == "texas") \%>\%
  st_extent()

class(tx_extent)

# Can use in joins
ngp \%>\%
  st_anti_join(tx_extent) \%>\%
  ggplot() +
  geom_sf(data = states_map) +
  geom_sf(data = tx_extent, fill = NA) +
  geom_sf(color = "red")

states_map \%>\%
  st_anti_join(tx_extent) \%>\%
  ggplot() +
  geom_sf() +
  geom_sf(data = tx_extent, fill = NA)

}
