% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-joins.R
\name{st_joins}
\alias{st_joins}
\alias{st_left_join}
\alias{st_inner_join}
\alias{st_semi_join}
\alias{st_anti_join}
\title{Join two 'sf' objects.}
\usage{
st_left_join(x, y, ...)

st_inner_join(x, y, ...)

st_semi_join(x, y, join = sf::st_intersects, ...)

st_anti_join(x, y, join = sf::st_intersects, ...)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{y}{object of class \code{sf}}

\item{...}{arguments passed on to the \code{join} function (e.g. \code{prepared}, or a pattern for \link{st_relate})}

\item{join}{geometry predicate function with the same profile as \link{st_intersects}; see details}
}
\description{
Join two 'sf' objects.
}
