% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-unitless.R
\name{unitless_measures}
\alias{unitless_measures}
\alias{st_ul_area}
\alias{st_ul_length}
\alias{st_ul_distance}
\title{Unitless measures}
\usage{
st_ul_area(x, ...)

st_ul_length(x)

st_ul_distance(
  x,
  y,
  ...,
  dist_fun,
  by_element = FALSE,
  which = ifelse(isTRUE(sf::st_is_longlat(x)), "Great Circle", "Euclidean"),
  par = 0,
  tolerance = 0
)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{...}{ignored}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg}, defaults to \code{x}}

\item{dist_fun}{deprecated}

\item{by_element}{logical; if \code{TRUE}, return a vector with distance between the first elements of \code{x} and \code{y}, the second, etc. if \code{FALSE}, return the dense matrix with all pairwise distances.}

\item{which}{character; for Cartesian coordinates only: one of \code{Euclidean}, \code{Hausdorff} or \code{Frechet}; for geodetic coordinates, great circle distances are computed; see details}

\item{par}{for \code{which} equal to \code{Hausdorff} or \code{Frechet}, optionally use a value between 0 and 1 to densify the geometry}

\item{tolerance}{ignored if \code{st_is_longlat(x)} is \code{FALSE}; otherwise, if set to a positive value, the first distance smaller than \code{tolerance} will be returned, and true distance may be smaller; this may speed up computation. In meters, or a \code{units} object convertible to meters.}
}
\description{
Unitless measures
}
\examples{
suppressPackageStartupMessages(library(dplyr))
data(states_map)

states_map \%>\%
  filter(sfx::st_ul_area(states_map) > 250e9)

}
